Move user state into auth state
