+++ frontend/src/business/models/Auth.ts
@@ -1,4 +1,7 @@
+import { User } from "./User";
+
 export interface AuthState {
   isSignedIn: boolean;
   token: string;
+  user: User | null;
 }
