+++ frontend/src/business/overmind/Auth/index.ts
@@ -1,3 +1,5 @@
+import * as actions from "./actions";
+import * as effects from "./effects";
 import { state } from "./state";
 
-export { state };
+export { actions, effects, state };
