+++ frontend/src/business/overmind/Auth/state.ts
@@ -3,4 +3,5 @@ import { AuthState } from "../../models";
 export const state: AuthState = {
   isSignedIn: false,
   token: "",
+  user: null,
 };
