--- frontend/src/business/overmind/User/index.ts
@@ -1,5 +1,0 @@
-import * as actions from "./actions";
-import * as effects from "./effects";
-import { state } from "./state";
-
-export { actions, effects, state };
