--- frontend/src/business/overmind/User/state.ts
@@ -1,5 +1,0 @@
-import { UserState } from "../../models";
-
-export const state: UserState = {
-  user: null,
-};
